+++ CHANGES.rst
@@ -4,6 +4,9 @@
 New Features
 ------------
 
+- New popout locations display Jdaviz in a detached popup window (``popout:window``)
+  or browser tab (``popout:tab``). [#1503]
+
 Cubeviz
 ^^^^^^^